# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sunflower Land Bot Constructor REST API
  description: |
    Вичерпна специфікація REST API серверного рушія Sunflower Land Bot Constructor.
    Забезпечує керування проектами, сесіями браузера Camoufox (антидетект Firefox),
    інвентарем ферми, статистикою та системними метриками.
  version: 1.0.0
  contact:
    name: Команда підтримки Sunflower Land Bot
servers:
  - url: http://localhost:3001
    description: Локальний сервер розробки
  - url: http://192.168.0.107:3001
    description: Локальна мережа (CasaOS / Home Server)
  - url: https://dxd.pp.ua
    description: Продакшн сервер з SSL
paths:
...
  /api/system/traces:
    get:
      summary: Останні транзакції та спани
      description: Повертає список останніх виконаних запитів із розбивкою на спани (Waterfall / Flamegraph).
      tags:
        - APM та Моніторинг
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Список транзакцій
          content:
            application/json:
              schema:
                type: object
  /api/system/slow-queries:
    get:
      summary: Список повільних SQL запитів
      description: Запити до SQLite бази даних, що виконувалися довше 20 мс.
      tags:
        - APM та Моніторинг
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Повільні запити
          content:
            application/json:
              schema:
                type: object
  /api/system/traces/otlp:
    get:
      summary: Експорт трас OpenTelemetry (OTLP JSON)
      description: Формує OpenTelemetry-сумісний потік трас для Jaeger, Zipkin, Tempo або Datadog.
      tags:
        - APM та Моніторинг
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OTLP ResourceSpans
          content:
            application/json:
              schema:
                type: object
  /api/system/rum:
    post:
      summary: Збереження RUM метрик з браузера клієнта
      description: Прийом Web Vitals (FCP, LCP, TTFB) та помилок JavaScript.
      tags:
        - APM та Моніторинг
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Метрики прийнято
    get:
      summary: Зведена статистика Real User Monitoring
      description: Отримання усереднених показників завантаження інтерфейсу користувачами.
      tags:
        - APM та Моніторинг
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Статистика RUM
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Передайте JWT токен авторизації у заголовку Authorization
    CSRFToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: CSRF токен для захисту від підробки запитів (отримується при handshake)
...